Make a more generic SQL store, tested against sqlite/mysql/postgres
The old pgxkv was a submodule, which is a pain. Also only PG. Make a more
generic SQL store, that doesn't require pgx directly - it has a stdlib shim
that should be OK. Also use this to expand to MySQL and SQLite support, with
generic support if it works. Easy to add dialects if needed.

Make a generic KV compliance test. Use this in a new submodule just to verify
the SQL KV against its desired targets. Run them in CI.