Move Opts around a little
Rather than requiring custom middleware types, move it to the context. This
lets us thread them through a normal set of middleware.

Introduce a way to define opts on a handler. This can be picked up in the
ServeHTTP method and passed via the context, so we can keep the mount time
configuration for Server