Move errors into own package, add Handler
This makes errors it's own thing. Adds a handler, that can intercept response
codes and pass them to the error handler. This will allow custom error pages,
without requiring the implementation to pass them in a special way. Panic
handling is added as well, now that we have a wrapper.