Use middleware chain, start making server less hidden
Ideally the server can be poked and prodded for users to customize as-is. Do
this by splitting out the middleware in to an exposed chain, and use different
muxes for the different middleware stack. We can then move to the setup taking
the bare minimum, and allow customization after construction.